ai-symptom-checker/requirements.txt
Flask==3.0.3
python-dotenv==0.21.0
google-generativeai==0.5.2
gunicorn==21.2.0          # For production deployment (optional)
requests==2.31.0          # If you're making any external API calls

